base impl for transfer. Need to implement callAndPoll with response tree parser

+++ icp_kotlin_kit/icp_candid/src/main/kotlin/com/bity/icp_candid/util/Timestamp.kt
@@ -6,7 +6,7 @@ import java.time.Instant
 
 fun icpTimestampNow(): CandidValue =
     Instant.now()
-        .epochSecond.
-        toULong()
+        .epochSecond
+        .toULong()
         .times(1_000_000_000UL)
         .icpTimestamp()
